+++ TechTestDDD.Domain/Entities/Vehicle.cs
@@ -3,9 +3,9 @@
     public class Vehicle
     {
         public int Id { get; set; }
-        public double Year { get; set;}
+        public float Year { get; set;}
         public string County { get; set; }
-        public double VehicleMiles { get; set; }
+        public float VehicleMiles { get; set; }
         public string Country { get; set; }
         public string State { get; set; }
     }
